Methods to control main menu visibility (in-progress, unstable)
Created GameObject variables for most items in scene. They need to be assigned, so the actual buttons and text for each needs to be placed in the scene then attached in the inspector. Then the methods need to be created for each view. See the paper I made to see what's in which view.

Removed visibility booleans, as they didn't really do anything.

I recommend looking into using lists to store the gameobjects in each view to simplify code significantly.